modified logic cart

+++ danny-store/src/main/java/danny/store/dannystore/domain/dto/CartDtoAndAmount.java
@@ -8,5 +8,5 @@ import java.util.List;
 @Data
 public class CartDtoAndAmount {
     private List<CartDto> cartDtoList;
-    private Long totalAmount;
+    private Float totalAmount;
 }
